+++ src/supermodels/adapters/__init__.py
@@ -1,0 +1,2 @@
+# ~/supermodels/src/supermodels/adapters/__init__.py
+from .sqla import SQLA, SQLAAdapter