+++ src/supermodels/adapters/sqla/__init__.py
@@ -1,0 +1,4 @@
+# ~/supermodels/src/supermodels/adapters/sqla/__init__.py
+
+from .adapter import SQLAAdapter
+SQLA = SQLAAdapter